Refactor runner retry tests
