+++ projects/04-llm-adapter/tests/test_runner_retry.py
@@ -1,295 +1,16 @@
-from pathlib import Path
-from types import SimpleNamespace
+from __future__ import annotations
 
-import pytest
+# NOTE: Legacy shim for runner retry tests.
+# Checklist:
+# - [ ] 04-llm-adapter tests import modules from projects/04-llm-adapter/tests/runner_retry
+# - [ ] Remove this shim once all references are updated
 
-from adapter.core.datasets import GoldenTask
-from adapter.core.errors import RateLimitError
-from adapter.core.metrics import BudgetSnapshot, RunMetrics
-from adapter.core.models import (
-    PricingConfig,
-    ProviderConfig,
-    QualityGatesConfig,
-    RateLimitConfig,
-    RetryConfig,
+from .runner_retry import test_rate_limit_failover as _test_rate_limit_failover
+from .runner_retry import test_shadow_metrics as _test_shadow_metrics
+from .runner_retry.test_rate_limit_failover import *  # noqa: F401,F403
+from .runner_retry.test_shadow_metrics import *  # noqa: F401,F403
+
+__all__ = (
+    _test_rate_limit_failover.__all__
+    + _test_shadow_metrics.__all__
 )
-from adapter.core.provider_spi import TokenUsage
-from adapter.core.providers import BaseProvider, ProviderResponse
-from adapter.core.runner_api import BackoffPolicy
-from adapter.core.runner_execution import RunnerExecution
-
-
-class _RateLimitStubProvider(BaseProvider):
-    def __init__(self, config: ProviderConfig, *, failures: int) -> None:
-        super().__init__(config)
-        self._failures = failures
-        self.calls = 0
-
-    def generate(self, prompt: str) -> ProviderResponse:
-        self.calls += 1
-        if self.calls <= self._failures:
-            raise RateLimitError("rate limited")
-        return ProviderResponse(
-            text="recovered",
-            latency_ms=5,
-            token_usage=TokenUsage(prompt=1, completion=1),
-        )
-
-
-class _SuccessProvider(BaseProvider):
-    def generate(self, prompt: str) -> ProviderResponse:
-        return ProviderResponse(
-            text="success",
-            latency_ms=3,
-            token_usage=TokenUsage(prompt=1, completion=1),
-        )
-
-
-class _TrackingProvider(BaseProvider):
-    def __init__(self, config: ProviderConfig, response: ProviderResponse) -> None:
-        super().__init__(config)
-        self._response = response
-        self.calls = 0
-
-    def generate(self, prompt: str) -> ProviderResponse:
-        self.calls += 1
-        return self._response
-
-
-class _UnusedProvider(BaseProvider):
-    def generate(self, prompt: str) -> ProviderResponse:  # pragma: no cover - defensive
-        raise AssertionError("unused provider should not be invoked")
-
-
-def _make_provider_config(tmp_path: Path, name: str, *, retries: RetryConfig | None = None) -> ProviderConfig:
-    retry_config = retries or RetryConfig()
-    return ProviderConfig(
-        path=tmp_path / f"{name}.yaml",
-        schema_version=1,
-        provider=name,
-        endpoint=None,
-        model=f"model-{name}",
-        auth_env=None,
-        seed=0,
-        temperature=0.0,
-        top_p=1.0,
-        max_tokens=10,
-        timeout_s=0,
-        retries=retry_config,
-        persist_output=True,
-        pricing=PricingConfig(),
-        rate_limit=RateLimitConfig(),
-        quality_gates=QualityGatesConfig(),
-        raw={},
-    )
-
-
-def _make_task() -> GoldenTask:
-    return GoldenTask(
-        task_id="task",
-        name="Task",
-        input={},
-        prompt_template="prompt",
-        expected={},
-    )
-
-
-def test_rate_limit_retry_advances_after_max(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
-    monkeypatch.setattr("adapter.core.runner_execution.sleep", lambda _seconds: None)
-    task = _make_task()
-    retry_config = RetryConfig(max=1, backoff_s=0.1)
-    failing_config = _make_provider_config(tmp_path, "rate-limit", retries=retry_config)
-    success_config = _make_provider_config(tmp_path, "next")
-    failing_provider = _RateLimitStubProvider(failing_config, failures=2)
-    success_provider = _SuccessProvider(success_config)
-
-    def evaluate_budget(
-        provider_config: ProviderConfig,
-        cost_usd: float,
-        status: str,
-        failure_kind: str | None,
-        error_message: str | None,
-    ) -> tuple[BudgetSnapshot, str | None, str, str | None, str | None]:
-        return BudgetSnapshot(0.0, False), None, status, failure_kind, error_message
-
-    def build_metrics(
-        provider_config: ProviderConfig,
-        task_obj: GoldenTask,
-        attempt_index: int,
-        mode: str,
-        response: ProviderResponse,
-        status: str,
-        failure_kind: str | None,
-        error_message: str | None,
-        latency_ms: int,
-        budget_snapshot: BudgetSnapshot,
-        cost_usd: float,
-    ) -> tuple[RunMetrics, str]:
-        metrics = RunMetrics(
-            ts="2024-01-01T00:00:00Z",
-            run_id=f"run-{provider_config.provider}",
-            provider=provider_config.provider,
-            model=provider_config.model,
-            mode=mode,
-            prompt_id=task_obj.task_id,
-            prompt_name=task_obj.name,
-            seed=provider_config.seed,
-            temperature=provider_config.temperature,
-            top_p=provider_config.top_p,
-            max_tokens=provider_config.max_tokens,
-            input_tokens=int(response.input_tokens),
-            output_tokens=int(response.output_tokens),
-            latency_ms=latency_ms,
-            cost_usd=cost_usd,
-            status=status,
-            failure_kind=failure_kind,
-            error_message=error_message,
-            output_text=response.output_text,
-            output_hash=None,
-            budget=budget_snapshot,
-        )
-        return metrics, response.output_text or ""
-
-    execution = RunnerExecution(
-        token_bucket=None,
-        schema_validator=None,
-        evaluate_budget=evaluate_budget,
-        build_metrics=build_metrics,
-        normalize_concurrency=lambda total, limit: total,
-        backoff=BackoffPolicy(timeout_next_provider=False, retryable_next_provider=False),
-        shadow_provider=None,
-        metrics_path=None,
-        provider_weights=None,
-    )
-
-    batch, stop_reason = execution.run_sequential_attempt(
-        [
-            (failing_config, failing_provider),
-            (success_config, success_provider),
-        ],
-        task,
-        attempt_index=0,
-        mode="sequential",
-    )
-
-    assert stop_reason is None
-    assert failing_provider.calls == 2
-    assert len(batch) == 2
-    first_result = batch[0][1]
-    assert first_result.metrics.status == "error"
-    assert first_result.metrics.retries == 1
-    assert first_result.backoff_next_provider is True
-    second_result = batch[1][1]
-    assert second_result.metrics.status == "ok"
-    assert second_result.metrics.retries == 0
-
-
-def test_run_sequential_attempt_updates_metrics_and_shadow(tmp_path: Path) -> None:
-    primary_config = _make_provider_config(tmp_path, "primary")
-    backup_config = _make_provider_config(tmp_path, "backup")
-    task = _make_task()
-    provider_response = ProviderResponse(
-        output_text="primary-response",
-        input_tokens=3,
-        output_tokens=2,
-        latency_ms=17,
-        token_usage=TokenUsage(prompt=3, completion=2),
-    )
-    primary_provider = _TrackingProvider(primary_config, provider_response)
-    backup_provider = _UnusedProvider(backup_config)
-    shadow_latency = 9
-
-    class _ShadowProvider:
-        def name(self) -> str:
-            return "shadow-test"
-
-        def capabilities(self) -> set[str]:
-            return set()
-
-        def invoke(self, request: object) -> SimpleNamespace:
-            return SimpleNamespace(latency_ms=shadow_latency)
-
-    def evaluate_budget(
-        provider_config: ProviderConfig,
-        cost_usd: float,
-        status: str,
-        failure_kind: str | None,
-        error_message: str | None,
-    ) -> tuple[BudgetSnapshot, str | None, str, str | None, str | None]:
-        return BudgetSnapshot(0.0, False), None, status, failure_kind, error_message
-
-    def build_metrics(
-        provider_config: ProviderConfig,
-        task_obj: GoldenTask,
-        attempt_index: int,
-        mode: str,
-        response: ProviderResponse,
-        status: str,
-        failure_kind: str | None,
-        error_message: str | None,
-        latency_ms: int,
-        budget_snapshot: BudgetSnapshot,
-        cost_usd: float,
-    ) -> tuple[RunMetrics, str]:
-        metrics = RunMetrics(
-            ts="2024-01-01T00:00:00Z",
-            run_id="run-primary",
-            provider=provider_config.provider,
-            model=provider_config.model,
-            mode=mode,
-            prompt_id=task_obj.task_id,
-            prompt_name=task_obj.name,
-            seed=provider_config.seed,
-            temperature=provider_config.temperature,
-            top_p=provider_config.top_p,
-            max_tokens=provider_config.max_tokens,
-            input_tokens=int(response.input_tokens),
-            output_tokens=int(response.output_tokens),
-            latency_ms=latency_ms,
-            cost_usd=cost_usd,
-            status=status,
-            failure_kind=failure_kind,
-            error_message=error_message,
-            output_text=response.output_text,
-            output_hash=None,
-            budget=budget_snapshot,
-        )
-        return metrics, response.output_text or ""
-
-    execution = RunnerExecution(
-        token_bucket=None,
-        schema_validator=None,
-        evaluate_budget=evaluate_budget,
-        build_metrics=build_metrics,
-        normalize_concurrency=lambda count, limit: count,
-        backoff=None,
-        shadow_provider=_ShadowProvider(),
-        metrics_path=None,
-        provider_weights=None,
-    )
-
-    batch, stop_reason = execution.run_sequential_attempt(
-        [
-            (primary_config, primary_provider),
-            (backup_config, backup_provider),
-        ],
-        task,
-        attempt_index=1,
-        mode="sequential",
-    )
-
-    assert stop_reason is None
-    assert len(batch) == 1
-    assert primary_provider.calls == 1
-
-    metrics = batch[0][1].metrics
-    assert metrics.providers == ["primary", "backup"]
-    assert metrics.token_usage == {"prompt": 3, "completion": 2, "total": 5}
-    assert metrics.attempts == 2
-    assert metrics.retries == 1
-    assert metrics.outcome == "success"
-    assert metrics.shadow_provider_id == "shadow-test"
-    assert metrics.shadow_latency_ms == shadow_latency
-    assert metrics.shadow_status == "ok"
-    assert metrics.shadow_outcome == "success"
-    assert metrics.shadow_error_message is None
